Implement build package github action
