+++ .github/workflows/publish_pypi.yml
@@ -1,0 +1,26 @@
+name: Build and publish package to PyPi
+
+on:
+  push:
+    branches: [ pypi-trigger ]
+
+
+jobs:
+  build_and_publish:
+
+    runs-on: ubuntu-latest
+    strategy:
+      matrix:
+        python-version: [3.7]
+
+    steps:
+      - uses: actions/checkout@v2
+      - name: Set up Python ${{ matrix.python-version }}
+        uses: actions/setup-python@v2
+        with:
+          python-version: ${{ matrix.python-version }}
+      - name: Build package
+        run: |
+          python -m pip install --upgrade pip
+          python -m pip install --upgrade setuptools wheel
+          python setup.py sdist bdist_wheel
