+++ extra_requirements.txt
@@ -1,4 +1,2 @@
 tpot==0.11.1
 h2o==3.28.1.2
-tensorflow==2.1.0; python_version <= '3.7'
-tensorflow==2.3.1; python_version == '3.8'
