+++ requirements.txt
@@ -17,3 +17,5 @@ pydot==1.4.1
 seaborn==0.9.0; python_version < '3.7'
 hyperopt==0.2.4
 joblib==0.17.0
+tensorflow==2.1.0; python_version <= '3.7'
+tensorflow==2.3.1; python_version == '3.8'
